Correccion de DomicilioService e impl

+++ src/main/java/projectBackend/Odontologia/Repository/DomicilioRepository.java
@@ -3,6 +3,5 @@ package projectBackend.Odontologia.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import projectBackend.Odontologia.Entity.Domicilio;
 
-public interface DomicilioRepository extends JpaRepository<Domicilio, Integer> {
-
+public interface DomicilioRepository extends JpaRepository<Domicilio, Long> {
 }
